fix atheme2json to handle mlock mode removal
For example, if the mlock is set to `-i`, `i` is not present in the
`modes` and thus it would error.

I'm inclined to think that the negative mlock feature doesn't behave
correctly, because the mlock of `-i` (or `-n`) would prevent anyone from
ever setting those modes on the channel. Which this does not appear to
be the case.

Fixes #1401